cleaning and refactoring auth
